Support -Dio.netty.recycler.maxCapacity.default but warn the user that it is deprecated.
Motivation:

In 4.0.x we used -Dio.netty.recycler.maxCapacity.default but in 4.1 we are using -Dio.netty.recycler.maxCapacity . We need to support both to not "break" users code.

Modifications:

Support the new and old system property but warn the user about the deprecation.

Result:

No breakage when upgrade from 4.0.x to 4.1.x.